Fixed software RAID failures in some circumstances
madm won't create a raid on a partition that already has a raid superblock.
Anaconda isn't very verbose about the reason for failure.

Added a pre-script that --zero-superblock on all partitions.
It may fail in case the pre-install partitions are different from the install ones, and those would correspond to ome previous raid one.
In this case, replay: this time the partitions are the ones that will be used, and zero will delete them.